Fix filter presenter in all viewmodels
