+++ Sources/CurrentFilters/ItemListController.swift
@@ -12,7 +12,7 @@ public protocol ItemListController: class {
 
   associatedtype Item: Hashable
 
-  func setItems(_ item: Set<Item>)
+  func setItems(_ item: [Item])
 
   // TODO: Potentially we could change from Item to a Int which is position of item in list.
   // It is enough to identify the items in viewModel, so in that way we only pass the
